@@ -37,32 +37,34 @@
     "compile-tests": "tsc -p e2e --watch"
   },
   "dependencies": {
-    "@angular/common": "~5.2.0",
-    "@angular/compiler": "~5.2.0",
-    "@angular/core": "~5.2.0",
-    "@angular/forms": "~5.2.0",
-    "@angular/http": "~5.2.0",
-    "@angular/platform-browser": "~5.2.0",
-    "@angular/platform-browser-dynamic": "~5.2.0",
-    "@angular/router": "~5.2.0",
+    "@angular/common": "~6.0.0-rc.3",
+    "@angular/compiler": "~6.0.0-rc.3",
+    "@angular/core": "~6.0.0-rc.3",
+    "@angular/forms": "~6.0.0-rc.3",
+    "@angular/http": "~6.0.0-rc.3",
+    "@angular/platform-browser": "~6.0.0-rc.3",
+    "@angular/platform-browser-dynamic": "~6.0.0-rc.3",
+    "@angular/router": "~6.0.0-rc.3",
     "email-validator": "^1.0.7",
-    "nativescript-angular": "~5.2.0",
+    "nativescript-angular": "next",
     "nativescript-iqkeyboardmanager": "~1.2.0",
     "nativescript-social-share": "~1.4.0",
     "nativescript-unit-test-runner": "^0.3.3",
     "reflect-metadata": "~0.1.8",
-    "rxjs": "^5.5.0",
+    "rxjs": "~6.0.0",
     "tns-core-modules": "4.0.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~5.2.0",
-    "@ngtools/webpack": "~1.9.1",
+    "@angular-devkit/core": "~0.5.5",
+    "@angular/compiler-cli": "~6.0.0-rc.3",
+    "@ngtools/webpack": "~6.0.0-rc.3",
     "babel-traverse": "6.24.1",
     "babel-types": "6.24.1",
     "babylon": "6.17.0",
+    "clean-webpack-plugin": "~0.1.19",
     "codelyzer": "^3.0.1",
-    "copy-webpack-plugin": "~4.3.0",
+    "copy-webpack-plugin": "~4.5.1",
     "css-loader": "~0.28.7",
     "extract-text-webpack-plugin": "~3.0.2",
     "filewalker": "^0.1.3",
@@ -72,17 +74,21 @@
     "karma-nativescript-launcher": "^0.4.0",
     "lazy": "1.0.11",
     "nativescript-css-loader": "~0.26.0",
-    "nativescript-dev-appium": "^3.2.0",
+    "nativescript-dev-appium": "~3.2.0",
     "nativescript-dev-typescript": "^0.6.0",
-    "nativescript-dev-webpack": "^0.9.1",
+    "nativescript-dev-webpack": "next",
     "nativescript-worker-loader": "~0.8.1",
     "raw-loader": "~0.5.1",
-    "resolve-url-loader": "~2.2.1",
+    "resolve-url-loader": "~2.3.0",
     "tslint": "^5.4.2",
-    "typescript": "~2.6.2",
-    "uglifyjs-webpack-plugin": "~1.1.6",
-    "webpack": "~3.10.0",
+    "typescript": "~2.7.2",
+    "uglifyjs-webpack-plugin": "~1.2.4",
+    "webpack": "~4.5.0",
     "webpack-bundle-analyzer": "^2.9.1",
-    "webpack-sources": "~1.1.0"
+    "webpack-cli": "~2.0.14",
+    "webpack-sources": "~1.1.0",
+    "@types/chai": "^4.0.2",
+    "@types/mocha": "^2.2.41",
+    "@types/node": "^7.0.5"
   }
 }
